backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
google-generativeai
requests
langchain
langchain-community
langchain-openai
